Modularized Account class
